Chore: Reorganized project file structure into assets and src folders
