+++ main.py
@@ -1,4 +1,4 @@
-from render import start
+from src.render import start
 
 if __name__ == "__main__":
     import sys
